+++ Editor/Artifice_Validator/Artifice_EditorWindow_Validator/Artifice_EditorWindow_Validator.cs
@@ -1,15 +1,14 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Artifice_Editor;
 using ArtificeToolkit.Editor.Resources;
 using ArtificeToolkit.Editor.VisualElements;
 using UnityEditor;
 using UnityEditor.SceneManagement;
 using UnityEditor.UIElements;
+using UnityEditorInternal;
 using UnityEngine;
 using UnityEngine.UIElements;
-using Object = UnityEngine.Object;
 
 // ReSharper disable InconsistentNaming
 
@@ -90,14 +89,14 @@ namespace ArtificeToolkit.Editor
             private readonly Label _objectNameLabel;
             private readonly Artifice_VisualElement_LabeledButton _autoFixButton;
 
-            private Object _originObject;
+            private Component _originComponent;
 
             public ValidatorLogListItem() : base()
             {
                 styleSheets.Add(Artifice_Utilities.GetGlobalStyle());
                 
                 // Create object label
-                _objectNameLabel = new Label("GameObject");
+                _objectNameLabel = new Label("Undefined");
                 _objectNameLabel.AddToClassList("object-name-label");
                 Add(_objectNameLabel);
                 
@@ -109,22 +108,31 @@ namespace ArtificeToolkit.Editor
                 RegisterCallback<ClickEvent>(evt =>
                 {
                     var listItem = (ValidatorLogListItem)evt.currentTarget;
-                    if(evt.clickCount == 2 && listItem._originObject != null)
-                       Selection.SetActiveObjectWithContext(listItem._originObject, listItem._originObject);
+                    if (evt.clickCount == 2 && listItem._originComponent != null)
+                    {
+                       Selection.SetActiveObjectWithContext(listItem._originComponent, listItem._originComponent);
+                       
+                       // Collapse all other components instead of our focused component.
+                       var components = _originComponent.gameObject.GetComponents<Component>();
+                       foreach (var component in components)
+                           InternalEditorUtility.SetIsInspectorExpanded(component, component == _originComponent);
+
+                       ActiveEditorTracker.sharedTracker.ForceRebuild();
+                    }
                 });
             }
 
             public void Set(Artifice_Validator.ValidatorLog log)
             {
-                base.Set(log.sprite, log.message);
-                _objectNameLabel.text = log.originObject == null ? "" : log.originObject.name;
-                _originObject = log.originObject;
+                base.Set(log.Sprite, log.Message);
+                _objectNameLabel.text = log.OriginComponent == null ? "" : log.OriginComponent.name;
+                _originComponent = log.OriginComponent;
 
-                if (log.hasAutoFix)
+                if (log.HasAutoFix)
                 {
                     // Show button and set callback
                     _autoFixButton.style.display = DisplayStyle.Flex;
-                    _autoFixButton.SetAction(log.autoFixAction);
+                    _autoFixButton.SetAction(log.AutoFixAction);
                 }
                 else
                     _autoFixButton.style.display = DisplayStyle.None;
@@ -176,7 +184,7 @@ namespace ArtificeToolkit.Editor
             
             // Initialize Filters
             _filters = new List<Func<Artifice_Validator.ValidatorLog, bool>>();
-            _filters.Add(log => OnSelectedScenesFilter(log) || OnSelectedAssetPathFilter(log));
+            // _filters.Add(log => OnSelectedScenesFilter(log) || OnSelectedAssetPathFilter(log));
             _filters.Add(OnSelectedValidatorTypesFilter);
             _filters.Add(OnLogTypeTogglesFilter);
             
@@ -200,14 +208,7 @@ namespace ArtificeToolkit.Editor
             _filteredLogs.Clear();
             foreach (var log in logs)
             {
-                // If a prefab stage is open, use prefab stage filter.
-                if (PrefabStageUtility.GetCurrentPrefabStage() != null)
-                {
-                    if(OnPrefabStageFilter(log))
-                        _filteredLogs.Add(log);
-                }
-                // Else, use all normal filters.
-                else if(_filters.All(filter => filter.Invoke(log)))
+                if(_filters.All(filter => filter.Invoke(log)))
                     _filteredLogs.Add(log);
             }
             _logsListView?.RefreshItems();
@@ -235,7 +236,7 @@ namespace ArtificeToolkit.Editor
             var trackedContainers = new ScrollView();
             trackedContainers.AddToClassList("tracked-containers-container");
             // trackedContainers.Add(BuildTrackedScenesUI()); // Removed for now. It seemed obnoxious and useless. 
-            trackedContainers.Add(BuildTrackedAssetFoldersUI());
+            // trackedContainers.Add(BuildTrackedAssetFoldersUI()); // Removed for now. It seems kind of useless with many validations being conflicting from asset space to scene space.
             trackedContainers.Add(BuildTrackedValidatorTypesUI());
             // Add to split-pane
             splitPane.Add(trackedContainers);
@@ -363,7 +364,7 @@ namespace ArtificeToolkit.Editor
             return container;
         }
         
-        private VisualElement BuildTrackedAssetFoldersUI()
+        private VisualElement BuildTrackedAssetFoldersUI() // Has been removed from the UI for now, but keep it for now in case we want to simply refactor the visuals in the future.
         {
             var container = new VisualElement();
             container.AddToClassList("tracked-list-container");
@@ -470,7 +471,7 @@ namespace ArtificeToolkit.Editor
                     var validatorTypeName = validatorModules[i].GetType().Name;
 
                     // Change display mode based on display on filters.
-                    if (validatorModules[i].DisplayOnFilters)
+                    if (validatorModules[i].DisplayOnFiltersList)
                         elem.style.display = DisplayStyle.Flex;
                     else
                         elem.style.display = DisplayStyle.None;
@@ -538,37 +539,30 @@ namespace ArtificeToolkit.Editor
 
         private bool OnSelectedScenesFilter(Artifice_Validator.ValidatorLog log)
         {
-            return _config.scenesMap.TryGetValue(log.originLocationName, out var value) && value
-                || log.originLocationName == "";
+            return _config.scenesMap.TryGetValue(log.OriginLocationName, out var value) && value
+                || log.OriginLocationName == "";
         }
 
         private bool OnSelectedValidatorTypesFilter(Artifice_Validator.ValidatorLog log)
         {
-            return _config.validatorTypesMap[log.originValidatorType.Name];
+            return _config.validatorTypesMap[log.OriginValidatorType.Name];
         }
 
         private bool OnLogTypeTogglesFilter(Artifice_Validator.ValidatorLog log)
         {
-            return _config.logTypesMap[log.logType];
+            return _config.logTypesMap[log.LogType];
         }
 
         private bool OnSelectedAssetPathFilter(Artifice_Validator.ValidatorLog log)
         {
             foreach (var (folderPath, shouldShow) in _config.assetPathsMap)
-                if (log.originLocationName.Contains(folderPath) && shouldShow)
+                if (log.OriginLocationName.Contains(folderPath) && shouldShow)
                     return true;
 
-            if (log.originLocationName == "")
+            if (log.OriginLocationName == "")
                 return true;
 
             return false;
-        }
-            
-        private bool OnPrefabStageFilter(Artifice_Validator.ValidatorLog log)
-        {
-            return log.originLocationName == PrefabStageKey &&
-                OnSelectedValidatorTypesFilter(log) &&
-                OnLogTypeTogglesFilter(log);
         }
         
         #endregion